+++ src/packages/v2-plugin/printers.ts
@@ -1,84 +1,20 @@
-import type { SubstitutePatch } from 'core-parts';
-import { makePatches, applyPatches } from 'core-parts';
-import type { AstPath, ParserOptions, Doc, Printer, Plugin } from 'prettier';
-import { format } from 'prettier';
-
-function sequentialFormattingAndTryMerging(options: ParserOptions, plugins: Plugin[]): string {
-  const { originalText } = options;
-  const sequentialFormattingOptions = {
-    ...options,
-    rangeEnd: Infinity,
-    plugins: [],
-  };
-
-  const firstFormattedText = format(originalText, sequentialFormattingOptions);
-
-  /**
-   * Changes that may be removed during the sequential formatting process.
-   */
-  const patches: SubstitutePatch[] = [];
-
-  const sequentiallyMergedText = plugins.reduce((formattedPrevText, plugin) => {
-    const temporaryFormattedText = format(formattedPrevText, {
-      ...sequentialFormattingOptions,
-      plugins: [plugin],
-    });
-
-    const temporaryFormattedTextWithoutPlugin = format(
-      temporaryFormattedText,
-      sequentialFormattingOptions,
-    );
-
-    patches.push(...makePatches(temporaryFormattedTextWithoutPlugin, temporaryFormattedText));
-
-    if (patches.length === 0) {
-      return temporaryFormattedText;
-    }
-
-    return applyPatches(temporaryFormattedTextWithoutPlugin, patches);
-  }, firstFormattedText);
-
-  return sequentiallyMergedText;
-}
+import type { AstPath, ParserOptions, Doc, Printer } from 'prettier';
 
 function createPrinter(): Printer {
   function main(
     path: AstPath,
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
     options: ParserOptions,
     // eslint-disable-next-line @typescript-eslint/no-unused-vars
     print: (path: AstPath) => Doc,
   ): Doc {
-    const plugins = options.plugins.filter((plugin) => typeof plugin !== 'string') as Plugin[];
-    const parserName = options.parser as string;
     const node = path.getValue();
 
-    if (node?.comments) {
-      node.comments.forEach((comment: any) => {
-        // eslint-disable-next-line no-param-reassign
-        comment.printed = true;
-      });
+    if (node.type === 'FormattedText') {
+      return node.body;
     }
 
-    const pluginIndex = plugins.findIndex(
-      (plugin) =>
-        Object.values(plugin.parsers ?? {}).every((parser) => parser.astFormat === 'merging-ast') &&
-        Object.entries(plugin.printers ?? {}).every(
-          ([astFormat, printer]) =>
-            astFormat === 'merging-ast' && Object.keys(printer).every((key) => key === 'print'),
-        ),
-    );
-
-    if (pluginIndex === -1) {
-      throw new Error(
-        "The structure of this plugin may have changed. If it's not in development, you may need to report an issue.",
-      );
-    }
-
-    const parserImplementedPlugins = plugins
-      .slice(0, pluginIndex)
-      .filter((plugin) => plugin.parsers?.[parserName]);
-
-    return sequentialFormattingAndTryMerging(options, parserImplementedPlugins);
+    throw new Error(`Unknown node type: ${node?.type}`);
   }
 
   return {
